+++ CHANGELOG.md
@@ -2,7 +2,20 @@
 
 ## Master
 
-- For expression now can contain `where` expression to filter array items. For example `{% for item in items where item > 1 %}` is now supported.
+### Enhancements
+
+- `for` block now can contain `where` expression to filter array items. For example `{% for item in items where item > 1 %}` is now supported.
+- `if` blocks may now contain else if (`elif`) conditions.
+
+  ```html+django
+  {% if one or two and not three %}
+    one or two but not three
+  {% elif four %}
+    four
+  {% else %}
+    not one, two, or four
+  {% endif %}
+  ```
 
 ## 0.8.0