+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Stencil Changelog
 
-## Master
+## 0.9.0
 
 ### Enhancements